Update Game.py
Implemented an early stage version of each space, so now each space has a location value and the players location depends on which space they are on. The spaces are held in a list. Values of each roll changes the space they are on rather than the position values themselves.